% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\docType{data}
\name{r}
\alias{r}
\title{Simulated 2D Weighted Data Set}
\format{
A list with 3 components:
\describe{
  \item{dat}{A data.frame of size \code{100000 x 2}, representing data point coordinates.}
  \item{m}{A numeric matrix of true peaks' coordinates.}
  \item{z}{A numeric vector of length \code{100000}, representing weights for each data point.}
}
}
\usage{
data(r)
}
\description{
This is a simulated dataset containing two-dimensional data points,
their corresponding weights, and the true peaks' coordinates.
}
\examples{
data(r)
}
\keyword{datasets}
